bases_tratadas/server/requirements.txt
djangorestframework
django-filter
django-cors-headers